moved TreeNode class into its own file
